\name{alreadyDone}

\alias{alreadyDone}

\title{alreadyDone}

\description{Checks if a doublet has already been added to a two-column matrix.}

\usage{
	out <- alreadyDone( donepairs , pair )
	}

\arguments{
	\item{pair}{A two-column matrix}
	\item{donepairs}{A doublet}
	}

\value{
	A boolean
	}

\examples{}