\name{comparerankings}

\alias{comparerankings}

\title{Compare two rankings}

\description{Counts the number of pairs of alternatives ranked differently in two rankings.}

\usage{
	out <- comparerankings( ranking1, ranking2 )
	}

\arguments{
	\item{ranking1}{A first ranking, i.e. a two-column matrix containing alternatives and their rank}
	\item{ranking1}{Another ranking, i.e. a two-column matrix containing alternatives and their rank}
	}

\value{An integer}

\examples{}