\name{computeAlternativesIndicator}

\alias{computeAlternativesIndicator}

\title{Computes alternatives indicator}

\description{Computes the necessary relation induced by the model}

\usage{
	value <- computeUtilitiesIndicator( M )
	}

\arguments{
	\item{M}{A model properly initialized and enriched with other functions of UTAR}
	}

\value{A real number}

\examples{}