\name{computeequation}

\alias{computeequation}

\title{Computes Hyperplane equation}

\description{Computes hyperplane equation from a pair of alternatives.}

\usage{
	out <- computeequation( a1, a2, M )
	}

\arguments{
	\item{a1}{An alternative, identified with its evaluation profile}
	\item{a2}{Another alternative, identified with its evaluation profile}
	\item{M}{A model properly initialized and enriched with other functions of UTAR}
	}

\value{}

\examples{}