\name{computeNecessaryRelation}

\alias{computeNecessaryRelation}

\title{Computes the necessary relation}

\description{Computes the necessary relation of a model on A.}

\usage{
	out <- computeNecessaryRelation( M, previousNecess )
	}

\arguments{
	\item{M}{A model properly initialized and enriched with other functions of UTAR}
	\item{previousNecess}{The necessary relation at the previous step of the procedure}
	}

\value{The updated necessary relation (set of pairs of alternatives, two-column matrix)}

\examples{}