\name{computeUtilitiesIndicator}

\alias{computeUtilitiesIndicator}

\title{computeUtilitiesIndicator}

\description{Computes and aggregates (product) the bounds of all variables of the model.}

\usage{
	value <- computeUtilitiesIndicator( M )
	}

\arguments{
	\item{M}{A model properly initialized and enriched with other functions of UTAR}
	}

\value{A real number}

\examples{}