\name{computeUtilitiesIndicator_weights}

\alias{computeUtilitiesIndicator_weights}

\title{computeUtilitiesIndicator_weights}

\description{Computes and aggregates (product) the bounds of trade-off weights variables of the model.}

\usage{
	value <- computeUtilitiesIndicator_weights( M )
	}

\arguments{
	\item{M}{A model properly initialized and enriched with other functions of UTAR}
	}

\value{A real number}

\examples{}