\name{countNbDiffAlternatives}

\alias{countNbDiffAlternatives}

\title{Counts the number of different alternatives}

\description{Counts the number of different alternatives in a doublet preference relation / indifference relation.}

\usage{
	n <- countNbDiffAlternatives( alternativesPreferences, alternativesIndifferences )
	}

\arguments{
	\item{alternativesPreferences}{A two-column matrix containing pairs of alternatives (preference statements)}
	\item{alternativesIndifferences}{A two-column matrix containing pairs of alternatives (indifference statements)}
	}

\value{
	An integer
	}

\examples{}