\name{extractSubsets}

\alias{extractSubsets}

\title{Constructs combinatory subsets}

\description{Constructs combinatory subsets (n-wise permutations).}

\usage{
	out <- extractSubsets( nA, n )
	}

\arguments{
	\item{nA}{Vector 1->nA}
	\item{n}{1<n<nA}
	}

\value{All n-wise permutations}

\examples{}