\name{findAndCheck}

\alias{findAndCheck}

\title{Finds and checks XML inputs and content}

\description{Finds and checks XML inputs and content in the inputsDirectory}

\usage{
	findAndCheck( name , inputsLocation , listOfFiles , aux )
	}

\arguments{
	\item{filename}{filename}
	\item{inputsLocation}{inputsLocation}
	\item{listOfFiles}{listOfFiles}
	\item{aux}{aux}
	}

\value{
	\item{result}{The appropriate result}
	\item{validation}{Validation tag (True if execution was successful)}
	\item{log}{If necessary, in case of error}
	}

\examples{}