\name{generateEasyLearningSet}

\alias{generateEasyLearningSet}

\title{Generates a set of easy pairs}

\description{Generates a set of easy pairs from a model M an its internal data.}

\usage{
	L <- generateEasyLearningSet( M )
	}

\arguments{
	\item{M}{A model properly initialized and enriched with the other functions of UTAR}
	}

\value{
	A matrix where each row is the concatenation of two evaluation profiles (of an easy pair)
	}

\examples{}