\name{generateRandomAlternatives}

\alias{generateRandomAlternatives}

\title{Generates a random set of alternatives}

\description{Generates a random set of alternatives in [0,1].}

\usage{
	out <- generateRandomAlternatives( n, p )
	}

\arguments{
	\item{n}{The number of criteria}
	\item{p}{The number of breakpoints in each dimension}
	}

\value{A matrix of n columns and p rows, containing the generated profiles}

\examples{}