\name{generateStandardAlternatives}

\alias{generateStandardAlternatives}

\title{Generates a regular grid of alternatives}

\description{Generates a regular grid of alternatives in [0,1].}

\usage{
	out <- generateStandardAlternatives( n, p )
	}

\arguments{
	\item{n}{The number of criteria}
	\item{p}{The number of breakpoints in each dimension}
	}

\value{A matrix of n columns and p^n rows, containing the generated profiles}

\examples{}