\name{produceOrientedSegment}

\alias{produceOrientedSegment}

\title{Produces oriented segment}

\description{Combines and computes oriented segment from criteria and segments information.}

\usage{
	updatedcontent <- produceOrientedSegment( content )
	}

\arguments{
	\item{content}{UTA initial data, extracted from \code{XML-XMCDA}, including criteria, alternatives, performance table and segments}
	}

\value{
	\item{updatedcontent}{Updated content}
	}

\examples{}