\name{restrictAllFilesToCorrectAlternatives}

\alias{restrictAllFilesToCorrectAlternatives}

\title{Restricts and combines content to correct alternatives}

\description{From all the initial data, computes the correct content, restricted to the selected alternatives.}

\usage{
	updatedcontent <- restrictAllFilesToCorrectAlternatives( content )
	}

\arguments{
	\item{content}{UTA initial data, extracted from XML-XMCDA, including criteria, alternatives, performance table and segments}
	}

\value{
	\item{updatedcontent}{Updated content}
	}

\examples{}