\name{restrictAllFilesToCorrectCriteria}

\alias{restrictAllFilesToCorrectCriteria}

\title{Restricts and combines content to correct criteria}

\description{From all the initial data, computes the correct content, restricted to the selected criteria.}

\usage{
	updatedcontent <- restrictAllFilesToCorrectCriteria( content )
	}

\arguments{
	\item{content}{UTA initial data, extracted from XML-XMCDA, including criteria, alternatives, performance table and segments}
	}

\value{
	\item{updatedcontent}{Updated content}
	}

\examples{}