\name{restrictPerformanceTable}

\alias{restrictPerformanceTable}

\title{Restricts and combines content to correct performance table}

\description{From all the initial data, computes the correct performance table, restricted to the selected alternatives and criteria.}

\usage{
	updatedcontent <- restrictPerformanceTable( content )
	}

\arguments{
	\item{content}{UTA initial data, extracted from XML-XMCDA, including criteria, alternatives, performance table and segments}
	}

\value{
	\item{updatedcontent}{Updated content}
	}

\examples{}