\name{saveUtilityFunctionUnderXML}

\alias{saveUtilityFunctionUnderXML}

\title{Exports value functions under XMCDA tree}

\description{Creates valid XMCDA tree for storing the value functions for gmatrix/umatrix structures}

\usage{
	saveUtilityFunctionUnderXML(valuefunction)
	}

\arguments{
	\item{valuefunction}{A list containing the gmatrix/umatrix matrices (value functions)}
	}

\value{
	Returns a valid XML-XMCDA tree
	}

\examples{}