\name{countNbDiffAlternatives}

\alias{countNbDiffAlternatives}

\title{Counts the number of different alternatives}

\description{Counts the number of different alternatives in a doublet preference relation / indifference relation}

\usage{
	countNbDiffAlternatives( alternativesPreferences, alternativesIndifferences )
	}

\arguments{
	\item{alternativesPreferences}{}
	\item{alternativesIndifferences}{}
	}

\value{
	An integer
	}

\examples{}