\name{getAnalyticCenterRanking}

\alias{getAnalyticCenterRanking}

\title{Computes analytic center ranking}

\description{Computes the ranking corresponding to the analytic center of the manipulated polyhedron.}

\usage{
	out <- getAnalyticCenterRanking( M )
	}

\arguments{
	\item{M}{A model properly initialized and enriched with other functions of UTAR}
	}

\value{A ranking, i.e. a two-column matrix containing alternatives and their rank}

\examples{}