\name{performanceTableToXML}

\alias{performanceTableToXML}

\title{Export performanceTable into XMCDA tree}

\description{Create an XML tree (according to XMCDA standard) based on performanceTable given as an input.}

\usage{
	z <- performanceTableToXML( performanceTable )
	}

\arguments{
	\item{data}{Classical performanceTable according to XMCDA standard. Matrix containing evaluations of alternatives on different criteria.}
	}

\value{}

\examples{}