\name{saveUtilityFunctionUnderXML}

\alias{saveUtilityFunctionUnderXML}

\title{Exports value functions under XML tree}

\description{Creates valid \code{XML-XMCDA} tree for storing the value functions for gmatrix/umatrix structures.}

\usage{
	out <- saveUtilityFunctionUnderXML(valuefunction)
	}

\arguments{
	\item{valuefunction}{A list containing the gmatrix/umatrix matrices (value functions)}
	}

\value{
	A valid \code{XML-XMCDA} tree
	}

\examples{}